+++ backend/src/proposals/proposals.module.ts
@@ -4,6 +4,8 @@ import { Flower } from '../flowers/flower.entity';
 import { NotificationsModule } from '../notifications/notifications.module';
 import { SharesModule } from '../shares/shares.module';
 import { SpeciesModule } from '../species/species.module';
+import { UsersModule } from '../users/users.module';
+import { ProfileStatsController } from './profile-stats.controller';
 import { ProposalsController } from './proposals.controller';
 import { ProposalsService } from './proposals.service';
 import { SpeciesProposal } from './species-proposal.entity';
@@ -14,8 +16,9 @@ import { SpeciesProposal } from './species-proposal.entity';
     SharesModule,
     NotificationsModule,
     SpeciesModule,
+    UsersModule,
   ],
-  controllers: [ProposalsController],
+  controllers: [ProposalsController, ProfileStatsController],
   providers: [ProposalsService],
 })
 export class ProposalsModule {}